[refactor/#224] tempEmailStore.ts 파일명 Zustand 네이밍 컨벤션 적용
