+++ src/queries/auth.ts
@@ -20,7 +20,7 @@ import {
   signup,
 } from "@/src/services/auth";
 import useOauthSignStore from "@/src/stores/oauthSignStore";
-import useSignupLinkStore from "@/src/stores/tempEmailStore";
+import useSignupLinkStore from "@/src/stores/useTempEmailStore";
 import useUserStore from "@/src/stores/userStore";
 import { OauthTypes } from "@/src/types/oauth";
 import {
